Refactor: move commands for custom-realm to right package
