+++ testsuite/standalone/src/test/java/org/wildfly/extras/creaper/commands/elytron/realm/AddCustomRealmImpl.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.wildfly.extras.creaper.commands.elytron.mapper;
+package org.wildfly.extras.creaper.commands.elytron.realm;
 
 import java.util.Map;
